requirements.txt
Flask==3.0.3
gunicorn==23.0.0
google-cloud-logging==3.11.2
google-cloud-firestore==2.18.0
